incorporate vector, host and activity parameters in the entomological model, initialize and build a model object

build_model_obj This function incorporates defined vector, host and activity parameters in the entomological model of mosquito oviposition cycle and builds a comprehensive object containing all model parameters. Additional parameters are calculated and included in the object. These include mosquito death and availability rate for three different types of hosts: humans protected by intervention, unprotected humans and animals). The function should be applied only on list objects created by the package functions def_vector_params and def_host_params. Functionality using other customized objects is not guaranteed.

build_model_obj(vec_p, hosts_p, activity, total_pop)

Arguments

vec_p

object containing vector-specific entomological parameteres created with the function def_vector_params

hosts_p

object containing host-specific entomological parameters created with the function def_host_params

total_pop

total population of humans and animals

Value

list object containing the following component:

vec_params: object with mosquito-specific entomological parameters (use function def_vector_params to generate object)

host_params: object with host-specific entomological parameters (use function def_hosts_params to generate object). In addition to previously-defined parameters, it contains:

muvA: mosquito death rate

alphai: availability to mosquitoes for each host type

activity: activity patterns for mosquitoes and humans (see function def_activity_patterns for object details)

total_pop: total population of humans and animals

Author

Monica Golumbeanu, monica.golumbeanu@swisstph.ch

Examples

# Retrieve the species-specific entomological parameters for An. gambiae
gambiae_vec_p = def_vector_params("Anopheles gambiae")
# Retrieve the host-specific entomological parameters for An. gambiae
gambiae_hosts_p = def_host_params("Anopheles gambiae")
# Retrieve default activity patterns for humans and mosquitoes
default_rhythms = def_activity_patterns("default_Anopheles_gambiae")
# Build model object for a population of 2000 hosts
model_p = build_model_obj(gambiae_vec_p, gambiae_hosts_p,
default_rhythms, 2000)
print(model_p)
#> $vec_params
#> $vec_params$species_name
#> [1] Anopheles gambiae
#> 75 Levels: Albitarsis Annularis Anopheles aconitus ... Sundaicus
#> 
#> $vec_params$M
#> [1] 0.6134499
#> 
#> $vec_params$M.sd
#> [1] 0.003819127
#> 
#> $vec_params$Chi
#> [1] 0.7947577
#> 
#> $vec_params$A0
#> [1] 0.6419328
#> 
#> $vec_params$A0.sd
#> [1] 0.07280429
#> 
#> $vec_params$zeta.3
#> [1] 1
#> 
#> $vec_params$td
#> [1] 0.33
#> 
#> $vec_params$tau
#> [1] 3
#> 
#> $vec_params$ts
#> [1] 10
#> 
#> $vec_params$to
#> [1] 5
#> 
#> $vec_params$endophily
#> [1] 0.7743085
#> 
#> $vec_params$endophily.sd
#> [1] 0.01233146
#> 
#> $vec_params$endophagy
#> [1] 0.5604133
#> 
#> $vec_params$endophagy.sd
#> [1] 0.00505659
#> 
#> 
#> $host_params
#> $host_params$species_name
#> [1] "Anopheles gambiae" "Anopheles gambiae"
#> 
#> $host_params$host
#> [1] "human"  "animal"
#> 
#> $host_params$PBi
#> [1] 0.95 0.95
#> 
#> $host_params$PCi
#> [1] 0.95 0.95
#> 
#> $host_params$PDi
#> [1] 0.99 0.99
#> 
#> $host_params$PEi
#> [1] 0.88 0.88
#> 
#> $host_params$Kvi
#> [1] 0.03 0.00
#> 
#> $host_params$muvA
#> [1] 0.6840223
#> 
#> $host_params$alphai
#> [1] 0.0019298339 0.0004983702
#> 
#> 
#> $activity
#> $activity$HBI
#>  [1] 0.000000000 0.000000000 0.000491400 0.000491400 0.005896806 0.007862408
#>  [7] 0.028009828 0.048157248 0.068304668 0.066339066 0.062899263 0.062899263
#> [13] 0.070270270 0.077641278 0.000000000 0.000000000
#> 
#> $activity$HBO
#>  [1] 0.000000000 0.000000000 0.000982801 0.000982801 0.003931204 0.025061425
#>  [7] 0.042260442 0.056019656 0.073710074 0.077149877 0.077149877 0.073710074
#> [13] 0.038329238 0.031449631 0.000000000 0.000000000
#> 
#> $activity$humans_indoors
#>  [1] 0.00000000 0.28380952 0.61904762 0.84761905 0.96190476 1.00000000
#>  [7] 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000 1.00000000
#> [13] 0.96190476 0.00952381 0.00000000 0.00000000
#> 
#> $activity$humans_in_bed
#>  [1] 0.0000000 0.0000000 0.0000000 0.2271333 0.6666333 1.0000000 1.0000000
#>  [8] 1.0000000 1.0000000 1.0000000 1.0000000 1.0000000 0.8833167 0.2006667
#> [15] 0.0000000 0.0000000
#> 
#> 
#> $total_pop
#> [1] 2000
#>