calculates the combined effects and impact for two interventions

calculate_combined_impact This function calculates the proportion by which vectorial capacity is reduced when a combination of two interventions is deployed. The user has the possibility to specify differing coverage for each intervention and the coverage of the combination.

calculate_combined_impact(
  combination_name = "combination",
  intervention1,
  intervention2,
  cov_intervention1,
  cov_intervention2,
  cov_combination = NULL,
  N_vec
)

Arguments

N_vec

number of daily emerging mosquitoes

combination_name:

name of the intervention combination

intervention1:

intervention object generated with def_intervection_effects()

intervention2:

intervention object generated with def_intervection_effects()

cov_intervention1:

coverage of intervention 1

cov_intervention2:

coverage of intervention 2

cov_combination:

coverage of the combination

Value

Author

Monica Golumbeanu, monica.golumbeanu@swisstph.ch

Tom Smith, thomas.smith@swisstph.ch

Clara Champagne, clara.champagne@swisstph.ch