calculates reduction in vectorial capacity for one interpolation point

calculate_impact_ip This function calculates the proportion by which vectorial capacity is reduced when an intervention is deployed to a cohort of humans at a given time point

calculate_impact_ip(intervention_obj, model_obj, n_vec, Ni)

Arguments

intervention_obj

object with the following attributes

PBi: probability that a mosquito bites host i

PCi: probability that a mosquito finds a resting place after biting a host of type i

PDi: probability that a mosquito survives the resting phase after biting a host of type i

PEi: probability that a mosquito lays eggs and returns to host-seeking after biting a host of type i

Kvi: proportion of susceptible mosquitoes that become infected after biting any host of type i

muvA: mosquito death rate

alphai: availability to mosquitoes for host of type i

model_obj

list of vector- and host-specific entomological parameters (see description of output from function calc_model_params)

Ni

data frame with the number of protected and unprotected hosts for various intervention coverages

N_vec

number of daily emerging mosquitoes

Value

object with the following attributes (columns = different coverages): #'

Pf_all: probability that mosquitoes survive the feeding cycle #'

or_all: delayed oocyst rate #'

sr_all: sporozoite rate #'

sigmai_all: host biting rate #'

eiri_all: entomological innoculation rate #'

vc_all: vectorial capacity #'

f_all: duration of the feeding cycle #'

PAmu_all: probability that mosquitoes die during host seeking #'

maxeig_all: maximum eigen value of the upsilon matrix

Author

Monica Golumbeanu, monica.golumbeanu@swisstph.ch

Olivier Briët, olivier.briet@swisstph.ch

Nakul Chitnis, nakul.chitnis@swisstph.ch

Tom Smith, thomas.smith@swisstph.ch