calculate intervention impact on the vectorial capacity and assess its variation accounting for the variation of input parameters

calc_impact_var calculates the proportion by which vectorial capacity is reduced for each of a list of interventions when deployed to a cohort of humans

calculate_impact_var(
  mosquito_species = "Anopheles gambiae",
  vec_ent_table = NULL,
  host_table = NULL,
  activity_patterns = "default_Anopheles_gambiae",
  interventions = intervention_obj_examples,
  coverage_vec = c(seq(0, 0.9, by = 0.1), 0.95, 0.99),
  total_pop = 2000,
  n_vec = 10000,
  n_time_points = 100,
  n_sample_points = 100,
  plot_result = TRUE
)

Arguments

mosquito_species

string corresponding to the name of the mosquito species to load the bionomic parameters for. This argument is case sensitive and must be one of the species provided with the package: Anopheles gambiae, Anopheles albimanus, etc. To see the available mosquito species in th package, you can use the function list_all_species(). Default value is Anopheles gambiae.

vec_ent_table

data frame with custom vector entomological values. Must be provided if custom parameter values should be used instead of the ones in the package database corresponding to the provided moquito species. The data frame should have the same structure (column names) as the database object vec_ent_param (see Data documentation). Default value is NULL.

host_table

data frame with custom host-specific values. Must be provided if custom parameter values should be used instead of the ones in the package database. The provided data frame should have the same structure as the database object host_ent_param (see Data documentation). Default value is NULL.

interventions

vector of intervention names

coverage_vec

vector of coverages to consider for each intervention

total_pop

total population size of human and animal hosts

n_vec

number of mosquitoes born and entering the host-seeking state per day

n_time_points

number of time points to be considered

n_sample_points

number of samples to select for assessing variance

plot_result

if TRUE, the function plots the reduction of the vectorial capacity for each intervention in the interventions list

activity

can be either a string corresponding to default entries: "default_Anopheles_gambiae", "default_Anopheles_albimanus" or a list object with the following attributes:

HBI: proportion of human biting indoors

HBO: propotion human biting outdoors

humans_indoors: proportion of humans indoor

humans_in_bed:proportion of humans in bed

These attribues can be either time series with the corresponding values at each time point, or characters indicating an entry ID in the package database. The function list_rhythms() can be used to retrieve the available entries for all geographical locations.

Value

list object with the following attributes:

vec_p: list of vector-specific entomological parameters (see description of output from function get_vector_params)

host_p: list of host-specific entomological parameters (see description of output from function get_host_params)

interventions_vec: vector of intervention-specific entomological parameters and impact (vectorial capacity) values

Author

Monica Golumbeanu, monica.golumbeanu@swisstph.ch