define host-specific entomological parameters

def_host_params This function can be used to load host-specific entomological parameters available in the package or to set the entomological parameters to new custom values provided by the user.

def_host_params(
  mosquito_species = "Anopheles gambiae",
  host_table = NULL,
  verbose = TRUE
)

Arguments

mosquito_species

string corresponding to the name of the mosquito species. Default value is "Anopheles gambiae". To be able to use the parameters from the package database, this name must match one of the Anopheles species provided in the package. To see all available Anopheles species, use the function list_all_species().

host_table

data frame with custom host-specific values. Must be provided if custom parameter values should be used instead of the ones in the package database. The provided data frame should have the same structure (column names) as the database object host_ent_param. Default value is NULL.

verbose

boolean specifying whether any messages should be displayed during the execution of the function. Default value is TRUE.

Value

list object with the following vector attributes:

species_name: name of the mosquito species

host: vector of hosts (should be = "human", "animal")

PBi: probability that a mosquito bites host i

PCi: probability that a mosquito finds a resting place after biting a host of type i

PDi: probability that a mosquito survives the resting phase after biting a host of type i

Kvi: proportion of susceptible mosquitoes that become infected after biting any host of type i

Each vector has three elements corresponding to the probability values for: (1) humans protected by interventions, (2) humans not protected by interventions, and (3) animals. In absence of interventions, the first two values of each vector are identical.

Author

Monica Golumbeanu, monica.golumbeanu@swisstph.ch

Olivier Briët

Tom Smith

Nakul Chitnis

Examples

# Retrieve the host-specific entomological parameters for An. gambiae
gambiae_hosts_p = def_host_params()

# Print all the parameters
print(gambiae_hosts_p)
#> $species_name
#> [1] "Anopheles gambiae" "Anopheles gambiae"
#> 
#> $host
#> [1] "human"  "animal"
#> 
#> $PBi
#> [1] 0.95 0.95
#> 
#> $PCi
#> [1] 0.95 0.95
#> 
#> $PDi
#> [1] 0.99 0.99
#> 
#> $PEi
#> [1] 0.88 0.88
#> 
#> $Kvi
#> [1] 0.03 0.00
#>