define the effects of interventions

This function defines the survival and effects of interventions on the transition probabilities between the consecutive stages of the mosquito oviposition cycle. These values are given for a set of interpolation points which are uniformly distributed across the duration of the intervention. To define the effects, this function either uses the intervention models which are available in the AnophelesModel database, or directly receives the vectors with the intervention effects for interventions which are not present in the database. In the latter case, it is assumed that the provided intervention effects values are adjusted according to the mosquito bionomics, activity patterns of mosquitoes and humans, as well as human exposure to mosquitoes.

def_interventions_effects(
  intervention_list,
  model_p,
  num_ip_points = 100,
  verbose = TRUE
)

Arguments

intervention_list

list of intervention objects, where each object is itself a list with the following attributes:

id: string with the intervention ID; for intervention IDs available in the package, check list_available_interventions()

description: short string with the description of the intervention; can be used to differentiate between interventions which have the same ID

parameterisation: string corresponding to the name of the intervention parameterisation to be used to calculate the transition probabilities. To use a parameterisation available in the package database, this attribute needs to match one of the parameterisations (column "Parameterisation") retrieved with list_intervention_models(). To use a new parameterisation which is not present in the package database, this argument can contain any name chosen by the user and the attribute effects needs to be provided as well (see description below).

LLIN_type: a string containing the LLIN type, only needed to be specified for LLINs interventions intending to use the LLINs models parameters provided with the package

LLIN_insecticide: a string containing the LLIN insecticide type, only needed for LLINs interventions intending to use the LLIN models parameters provided with the package,

LLIN_country: a string containing the country where the LLIN characteristics were measured, only provided for LLINs interventions intending to use the LLIN models parameters provided with the package

effects: list of intervention survival and effects on the mosquito oviposition cycle. This attribute allows defining intervention effects which are not in the package database and is optional if parameterisation matches one of the intervention parameterisations available in the package database. These effects include, in addition to the availability to mosquitoes and the proportion of infected mosquitoes, the transition probabilities between the consecutive stages of the oviposition cycle. The decay of the effects can also be specified if generation of a <GVI> xml snippet for OpenMalaria simulations is envisaged. The effects are defined through the following list attributes:

host_types: for reference only, vector of strings coding the different types of hosts, must be equal to {"protected humans", "unprotected humans", "animals"}. All the remaining attributes of the effects object, except for survival are represented by 3-column matrices, where each column, j={1, 2, 3}, corresponds to values for the three types of hosts defined in the same order as for host_types

Kvi: proportion of susceptible mosquitoes that become infected after biting a host of type j, contains one row and 3 columns

alphai: availability of a host of type j to mosquitoes, contains num_ip_points rows and 3 columns

PBi: probability that a mosquito bites a host of type j, contains num_ip_points rows and 3 columns

PCi: probability that a mosquito finds a resting place after biting a host of type j, contains num_ip_points rows and 3 columns

PDi: probability that a mosquito survives the resting phase after biting a host of type j, contains num_ip_points rows and 3 columns

PEi: probability that a mosquito lays eggs and returns to host-seeking after biting a host of type j, contains num_ip_points rows and 3 columns

survival: survival of the intervention, vector of length num_ip_points

alphai_decay: decay of intervention effect on host availability to mosquitoes, contains num_ip_points rows and 3 columns

PBi_decay: decay of intervention effect on the probability that a mosquito bites a host of type j, contains num_ip_points rows and 3 columns

PCi_decay: decay of intervention effect on the probability that a mosquito finds a resting place after biting a host of type j, contains num_ip_points rows and 3 columns

For the above attributes with num_ip_points rows, the values are provided for consecutive time points equally distributed throughout the duration of the intervention. The number of time points needs to match the provided argument n_time_points.

model_p

model object (generated with function build_model_obj())

num_ip_points

number of interpolation points, default is 100. If custom intervention effects are specified, the relevant effect attributes need to have a number of rows equal to num_ip_points (see above)

verbose

boolean specifying whether any messages should be displayed during the execution of the function. Default value is TRUE.

Value

list object with the following attributes:

id: string with the intervention ID

description: short string with the description of the intervention

parameterisation: string corresponding to the name of the intervention parameterisation

LLIN_type: only for LLINs, a string containing the LLIN type

LLIN_insecticide: only for LLINs, a string containing the LLIN insecticide type

LLIN_country: only for LLINs, a string containing the country where the LLIN characteristics were measured

effects: List of intervention survival and effects on the mosquito oviposition cycle (same format as defined above for the input attributes)

Author

Monica Golumbeanu, monica.golumbeanu@swisstph.ch