define the vector-specific entomological parameters

def_vector_params This function can be used to load entomological parameters for mosquito species available in the package database or to set the entomological parameters to new custom values provided by the user.

def_vector_params(
  mosquito_species = "Anopheles gambiae",
  vector_table = NULL,
  verbose = TRUE
)

Arguments

mosquito_species

string corresponding to the name of the mosquito species. Default value is "Anopheles gambiae". To be able to use the parameters from the package database, this name must match one of the Anopheles species provided in the package. To see all available Anopheles species, use the function list_all_species().

vector_table

data frame with custom vector entomological values. Must be provided if custom parameter values should be used instead of the ones in the package database. The data frame should have the same structure (column names) as the database object vec_ent_param. Default value is NULL.

verbose

boolean specifying whether any messages should be displayed during the execution of the function. Default value is TRUE.

Value

list object with the following attributes (same as vec_ent_param):

species_name: name of the mosquito species

M: parous rate (proportion of host seeking mosquitoes that have laid eggs at least once) #'

M.sd: standard deviation of the parous rate

Chi: human blood index (proportion of mosquito blood meals derived from humans)

A0: sac rate (proportion of mosquitoes who laid eggs in a day)

A0.sd: standard deviation of the sac rate

zeta.3: relative availability of different non-human hosts

td: proportion of a day that a mosquito actively seeks a host

tau: time required for a mosquito that has encountered a host to return to host seeking

ts: duration of the extrinsic incubation period (time required for sporozoites to develop in mosquitoes)

endophily: proportion of indoor resting mosquitoes

endophily.sd: standard deviation of the endophily

endophagy: proportion of indoor feeding mosquitoes

endophagy.sd: standard deviation of the endophagy

Author

Monica Golumbeanu, monica.golumbeanu@swisstph.ch

Examples

# Retrieve the vector entomological parameters for An. gambiae
gambiae_vec_p = def_vector_params()

# Print all the parameters
print(gambiae_vec_p)
#> $species_name
#> [1] Anopheles gambiae
#> 75 Levels: Albitarsis Annularis Anopheles aconitus ... Sundaicus
#> 
#> $M
#> [1] 0.6134499
#> 
#> $M.sd
#> [1] 0.003819127
#> 
#> $Chi
#> [1] 0.7947577
#> 
#> $A0
#> [1] 0.6419328
#> 
#> $A0.sd
#> [1] 0.07280429
#> 
#> $zeta.3
#> [1] 1
#> 
#> $td
#> [1] 0.33
#> 
#> $tau
#> [1] 3
#> 
#> $ts
#> [1] 10
#> 
#> $to
#> [1] 5
#> 
#> $endophily
#> [1] 0.7743085
#> 
#> $endophily.sd
#> [1] 0.01233146
#> 
#> $endophagy
#> [1] 0.5604133
#> 
#> $endophagy.sd
#> [1] 0.00505659
#>