creates an xml snippet to be included in an OpenMalaria scenario for the GVI intervention

get_OM_GVI_snippet This function can be used to create the relevant xml tags for the <GVI> component of the scenario xml file required as input for running the OpenMalaria agent-based simulation model. Precisely, it creates the <mosq> element containing parameters describing the mosquito life cycle, as well as the <nonHumanHosts> element with parameters specific to animal hosts.

get_OM_GVI_snippet(species, intervention_impact, num_points, plot_f = TRUE)

Arguments

intervention_impact

object created by the AnophelesModel main function

Value

TO DO

References

TO DO

Author

Monica Golumbeanu, monica.golumbeanu@swisstph.ch

Olivier Briët, olivier.briet@swisstph.ch

Nakul Chitnis, nakul.chitnis@swisstph.ch

Tom Smith, thomas.smith@swisstph.ch