creates an xml snippet to be included in an OpenMalaria scenario

get_OM_ento_snippet Once the vector and host parameters defined, this function can be used to generate the xml tags for the <entomology> component of the scenario xml file required as input for running the OpenMalaria agent-based simulation model of malaria epidemiology. It creates the <mosq> element containing parameters describing the mosquito life cycle, as well as the <nonHumanHosts> element with parameters for non-human hosts.

get_OM_ento_snippet(vec_params, hosts_params)

Arguments

vec_p

list of vector-specific entomological parameteres (see def_vector_params for creating this object)

hosts_p

list of host-specific entomological parameters (see def_host_params for creating this object)

Value

list with two xml snippets:

<mosq>: mosquito species parameters describing the lifecycle and behaviour

<nonHumanHosts>: non-human parameters to allow for zoophily

Author

Monica Golumbeanu, monica.golumbeanu@swisstph.ch

Examples

vec_params = def_vector_params()
hosts_params = def_host_params()
xml_snippets = get_OM_ento_snippet(vec_params, hosts_params)
print(xml_snippets)
#> $mosq_snippet
#> <mosq minInfectedThreshold="0.01">
#>   <mosqRestDuration value="2"/>
#>   <extrinsicIncubationPeriod value="10"/>
#>   <mosqLaidEggsSameDayProportion value="0.641932826304081"/>
#>   <mosqSeekingDuration value="3"/>
#>   <mosqSurvivalFeedingCycleProbability value="0.613449878545352"/>
#>   <availability/>
#>   <mosqProbBiting mean="0.95" variance="0"/>
#>   <mosqProbFindRestSite mean="0.95" variance="0"/>
#>   <mosqProbResting mean="0.99" variance="0"/>
#>   <mosqProbOvipositing mean="0.88"/>
#>   <mosqHumanBloodIndex mean="0.794757704738354"/>
#> </mosq> 
#> 
#> $nonHumanHosts_snippet
#> <nonHumanHosts name="unprotectedAnimals">
#>   <mosqRelativeEntoAvailability value="1"/>
#>   <mosqProbBiting value="0.95"/>
#>   <mosqProbFindRestSite value="0.95"/>
#>   <mosqProbResting value="0.99"/>
#> </nonHumanHosts> 
#>