Calculate human exposure to biting

get_in_out_exp This function can be used to calculate indoor and outdoor exposure to biting.

get_in_out_exp(activity_cycles, vec_p)

Arguments

activity_cycles

data frame with the following columns:

HBI: proportion of human biting indoors

HBO: propotion human biting outdoors

humans_indoors: proportion of humans indoors

humans_in_bed:proportion of humans in bed

vec_p

list of vector-specific entomological paramteres (see description of output from function def_vector_params)

Value

The function returns an object with the following attributes:

Exposure_Indoor_total: indoor exposure proportion

Exposure_Outdoor_total: outdoor exposure proportion

Exposure_Indoor_whileinbed: indoor in bed exposure proportion

Exposure_Outdoor_whileinbed: outdoor in bed exposure proportion

indoor_resting: proportion of mosquitoes that rest indoors.

All numbers in the output list are proportions of the overall total exposure in the absence of interventions.

References

TO DO

Author

Monica Golumbeanu, monica.golumbeanu@swisstph.ch

Olivier Briët, olivier.briet@swisstph.ch

Nakul Chitnis, nakul.chitnis@swisstph.ch

Tom Smith, thomas.smith@swisstph.ch