Host-specific estimated parameters for the entomological model of malaria transmission

This data object contains for each of the three host types (humans protected by interventions, non-protected humans, and animals) estimated parameters for the entomological model of malaria transmission.

host_ent_param

Format

A data frame with 5 columns consisting of:

species name: name of the mosquito species

probability: probabilities that a mosquito completes the next stage of the oviposition cycle in one night, conditional on having reached that stage. Can only be one of the following: PBi(probability that a mosquito bites host of type i), PCi(probability that a mosquito finds a resting place after biting a host of type i), PDi (probability that a mosquito survives the resting phase after biting a host of type i), PEi(probability that a mosquito lays eggs and returns to host-seeking after biting a host of type i), Kvi(proportion of susceptible mosquitoes that become infected after biting a host of type i)

host_type_1: value of the probability for host type 1 (humans protected by the intervention)

host_type_2: value of the probability for host type 2 (humans not protected by any intervention)

host_type_3: value of the probability for host type 3 (animals)

Value

host-specific entomological parameters

References

TO DO

Author

Monica Golumbeanu, monica.golumbeanu@swisstph.ch

Olivier Briët, olivier.briet@swisstph.ch

Nakul Chitnis, nakul.chitnis@swisstph.ch

Tom Smith, thomas.smith@swisstph.ch

Examples

# Print the first lines of the data frame object
head(host_ent_param)
#>        species_name   host  PBi  PCi  PDi  PEi  Kvi
#> 1 Anopheles gambiae  human 0.95 0.95 0.99 0.88 0.03
#> 2 Anopheles gambiae animal 0.95 0.95 0.99 0.88 0.00