General description of IRS and LLINs interventions parameterizations

This data object contains a general description of all available IRS and LLIN intervention parameterizations.

interventions_param

Format

A data frame with 7 columns consisting of:

Parameterisation: ID of the parameterization

Intervention: name identified of the intervention, can only take value "IRS", "LLINs" or "House screening"

Active agent: insecticide name

Species: name of the mosquito species

Reference: publication reference

Method: fitting method

Duration: duration until intervention renewal

Value

general intervention description

Author

Monica Golumbeanu, monica.golumbeanu@swisstph.ch

Olivier Briët,

Nakul Chitnis, nakul.chitnis@swisstph.ch

Tom Smith, thomas.smith@unibas.ch

Examples

# Print the first lines of the data frame object
head(interventions_param)
#> $interventions_summary
#>    Parameterisation    Intervention       Active_agent                Species
#> 1             IRS01             IRS                DDT    Anopheles albimanus
#> 2             IRS10             IRS               DDT  Anopheles gambiae s.l.
#> 3             IRS12             IRS                DDT Anopheles gambiae s.s.
#> 4             IRS14             IRS   Deltamethrin IRS Anopheles gambiae s.s.
#> 5             IRS16             IRS     Permethrin IRS Anopheles gambiae s.l.
#> 6             IRS02             IRS         Icon 10 CS Anopheles gambiae s.s.
#> 7             IRS03             IRS   Actellic 300 CS  Anopheles gambiae s.s.
#> 8             IRS04             IRS     Actellic 50 EC Anopheles gambiae s.s.
#> 9             IRS05             IRS        Bendiocarb  Anopheles gambiae s.l.
#> 10            IRS06             IRS        Bendiocarb     Anopheles albimanus
#> 11            IRS08             IRS        Bendiocarb  Anopheles gambiae s.s.
#> 12            IRS17             IRS Lambdacyhalothrin  Anopheles gambiae s.s.
#> 13          LLINs01           LLINs          PN2 Akron      Anopheles gambiae
#> 14          LLINs02           LLINs         PN2 Zeneti      Anopheles gambiae
#> 15          LLINs03           LLINs         Malanville      Anopheles gambiae
#> 16          LLINs04           LLINs Lambdacyhalothrin     Anopheles albimanus
#> 17      Screening01 House_screening               <NA>      Anopheles gambiae
#>                         Reference Duration
#> 1                  Bangs (thesis)      0.5
#> 2                    Chitnis 2008      1.0
#> 3                     Kuhlow 1962      0.5
#> 4                Bown et al. 1987      0.5
#> 5                    Chitnis 2008      1.0
#> 6                   Tchicaya 2014      1.0
#> 7                   Tchicaya 2014      1.0
#> 8                   Tchicaya 2014      1.0
#> 9                    Chitnis 2008      1.0
#> 10               Bown et al. 1987      0.5
#> 11                    Agossa 2014      0.5
#> 12                    Agossa 2014      0.5
#> 13 Randriamaherijaona et al. 2015      3.0
#> 14 Randriamaherijaona et al. 2015      3.0
#> 15 Randriamaherijaona et al. 2015      3.0
#> 16 Randriamaherijaona et al. 2015      3.0
#> 17                           <NA>      3.0
#> 
#> $LLINs_params
#> $LLINs_params$model_coeff
#>         beta0.PEnt beta1.PEnt beta0.PAtt beta1.PAtt beta2.PAtt beta3.PAtt
#> LLINs01    4.59512    -0.9860   -4.06900    0.54100    0.80700   -0.10100
#> LLINs02    4.59512    -0.8440   -2.55500    0.37900    1.80700   -0.07800
#> LLINs03    4.59512    -1.2120   -0.02000    0.22900    2.00600   -0.17300
#> LLINs04    4.59512    -0.5340   -0.86100   -0.00500    0.25300   -0.02100
#> LLINs05    4.08680    -2.8851   -0.75405    0.47219    1.47636   -0.17521
#>         beta0.PBmu beta1.PBmu beta2.PBmu beta3.PBmu beta0.PCmu beta1.PCmu
#> LLINs01    -4.8400   -0.12800    2.32500   -0.11300  -4.595120  -0.114000
#> LLINs02    -1.8230   -0.23500    2.24500   -0.09400  -4.595120  -0.370000
#> LLINs03    -2.9140   -0.31300    3.86400   -0.18400  -4.595120   0.046000
#> LLINs04    -1.3700   -0.02500    0.89300   -0.03500  -2.501000  -0.021000
#> LLINs05    -3.0083   -0.34248    3.44955   -0.08499  -3.800727  -0.011861
#>         beta2.PCmu beta3.PCmu HolesMax
#> LLINs01   1.303000   0.000000   192000
#> LLINs02   2.816000   0.000000   192000
#> LLINs03   1.470000   0.000000   192000
#> LLINs04   0.683000  -0.031000   192000
#> LLINs05   3.231888   0.008622   145200
#> 
#> $LLINs_params$insecticide_c
#>                    Default PermaNet 2.0 DawaPlus 2.0    DuraNet Interceptor
#> initialInsecticide      55           55   80.0000000 261.000000 200.0000000
#> scalingfactor            1            1    0.9952966   1.042892   0.6822763
#>                       LifeNet Netprotect       Olyset PermaNet 3.0
#> initialInsecticide 63.0000000 68.1000000 1000.0000000     85.00000
#> scalingfactor       0.9907015  0.8310243    0.3549849      2.31743
#>                    PermaNet 3.0 lower side panels
#> initialInsecticide                       85.00000
#> scalingfactor                             2.31743
#>                    PermaNet 3.0 upper side panels PermaNet 3.0 top panel
#> initialInsecticide                      85.000000             121.000000
#> scalingfactor                            2.091642               1.019788
#>                    Lambdacyhalothrin (gambiae) Lambdacyhalothrin (albimanus)
#> initialInsecticide                          18                            30
#> scalingfactor                                1                             1
#> 
#> $LLINs_params$durability_estim
#>                              Parameter     Net_Type    Country Semester Value
#> 1   Mean of log transformed holed area      DuraNet      Kenya        1   0.1
#> 2     SD of log transformed holed area      DuraNet      Kenya        1   0.4
#> 3                             Survival      DuraNet      Kenya        1 100.0
#> 4   Mean of log transformed holed area      DuraNet      Kenya        2   1.1
#> 5     SD of log transformed holed area      DuraNet      Kenya        2   2.0
#> 6                             Survival      DuraNet      Kenya        2  99.6
#> 7   Mean of log transformed holed area      DuraNet      Kenya        3   2.0
#> 8     SD of log transformed holed area      DuraNet      Kenya        3   2.0
#> 9                             Survival      DuraNet      Kenya        3  99.6
#> 10  Mean of log transformed holed area      DuraNet      Kenya        4   1.2
#> 11    SD of log transformed holed area      DuraNet      Kenya        4   1.3
#> 12                            Survival      DuraNet      Kenya        4  97.1
#> 13  Mean of log transformed holed area      DuraNet      Kenya        5   1.4
#> 14    SD of log transformed holed area      DuraNet      Kenya        5   1.9
#> 15                            Survival      DuraNet      Kenya        5  97.1
#> 16  Mean of log transformed holed area      DuraNet      Kenya        6   2.5
#> 17    SD of log transformed holed area      DuraNet      Kenya        6   2.6
#> 18                            Survival      DuraNet      Kenya        6  96.3
#> 19  Mean of log transformed holed area      DuraNet      Kenya        7   3.4
#> 20    SD of log transformed holed area      DuraNet      Kenya        7   2.3
#> 21                            Survival      DuraNet      Kenya        7  94.4
#> 22  Mean of log transformed holed area      DuraNet      Kenya        8   3.3
#> 23    SD of log transformed holed area      DuraNet      Kenya        8   2.8
#> 24                            Survival      DuraNet      Kenya        8  89.3
#> 25  Mean of log transformed holed area      DuraNet     Malawi        1   1.1
#> 26    SD of log transformed holed area      DuraNet     Malawi        1   2.0
#> 27                            Survival      DuraNet     Malawi        1 100.0
#> 28  Mean of log transformed holed area      DuraNet     Malawi        2   1.7
#> 29    SD of log transformed holed area      DuraNet     Malawi        2   1.4
#> 30                            Survival      DuraNet     Malawi        2  98.4
#> 31  Mean of log transformed holed area      DuraNet     Malawi        3   1.9
#> 32    SD of log transformed holed area      DuraNet     Malawi        3   1.8
#> 33                            Survival      DuraNet     Malawi        3  96.4
#> 34  Mean of log transformed holed area      DuraNet     Malawi        4   2.3
#> 35    SD of log transformed holed area      DuraNet     Malawi        4   1.7
#> 36                            Survival      DuraNet     Malawi        4  95.5
#> 37  Mean of log transformed holed area      DuraNet     Malawi        5   2.6
#> 38    SD of log transformed holed area      DuraNet     Malawi        5   1.8
#> 39                            Survival      DuraNet     Malawi        5  87.8
#> 40  Mean of log transformed holed area      DuraNet     Malawi        6   3.3
#> 41    SD of log transformed holed area      DuraNet     Malawi        6   1.7
#> 42                            Survival      DuraNet     Malawi        6  65.0
#> 43  Mean of log transformed holed area  Interceptor      Kenya        1   1.4
#> 44    SD of log transformed holed area  Interceptor      Kenya        1   2.0
#> 45                            Survival  Interceptor      Kenya        1 100.0
#> 46  Mean of log transformed holed area  Interceptor      Kenya        2   2.3
#> 47    SD of log transformed holed area  Interceptor      Kenya        2   2.8
#> 48                            Survival  Interceptor      Kenya        2  99.5
#> 49  Mean of log transformed holed area  Interceptor      Kenya        3   2.0
#> 50    SD of log transformed holed area  Interceptor      Kenya        3   2.2
#> 51                            Survival  Interceptor      Kenya        3  99.5
#> 52  Mean of log transformed holed area  Interceptor      Kenya        4   2.4
#> 53    SD of log transformed holed area  Interceptor      Kenya        4   2.2
#> 54                            Survival  Interceptor      Kenya        4  99.1
#> 55  Mean of log transformed holed area  Interceptor      Kenya        5   3.1
#> 56    SD of log transformed holed area  Interceptor      Kenya        5   2.6
#> 57                            Survival  Interceptor      Kenya        5  97.7
#> 58  Mean of log transformed holed area  Interceptor      Kenya        6   3.6
#> 59    SD of log transformed holed area  Interceptor      Kenya        6   2.7
#> 60                            Survival  Interceptor      Kenya        6  91.4
#> 61  Mean of log transformed holed area  Interceptor      Kenya        7   5.7
#> 62    SD of log transformed holed area  Interceptor      Kenya        7   3.8
#> 63                            Survival  Interceptor      Kenya        7  72.4
#> 64  Mean of log transformed holed area  Interceptor      Kenya        8   4.8
#> 65    SD of log transformed holed area  Interceptor      Kenya        8   3.4
#> 66                            Survival  Interceptor      Kenya        8  50.4
#> 67  Mean of log transformed holed area  Interceptor     Malawi        1   0.4
#> 68    SD of log transformed holed area  Interceptor     Malawi        1   1.2
#> 69                            Survival  Interceptor     Malawi        1  99.9
#> 70  Mean of log transformed holed area  Interceptor     Malawi        2   2.1
#> 71    SD of log transformed holed area  Interceptor     Malawi        2   2.0
#> 72                            Survival  Interceptor     Malawi        2  99.4
#> 73  Mean of log transformed holed area  Interceptor     Malawi        3   3.2
#> 74    SD of log transformed holed area  Interceptor     Malawi        3   2.6
#> 75                            Survival  Interceptor     Malawi        3  95.3
#> 76  Mean of log transformed holed area  Interceptor     Malawi        4   3.4
#> 77    SD of log transformed holed area  Interceptor     Malawi        4   2.0
#> 78                            Survival  Interceptor     Malawi        4  88.9
#> 79  Mean of log transformed holed area  Interceptor     Malawi        5   4.6
#> 80    SD of log transformed holed area  Interceptor     Malawi        5   2.1
#> 81                            Survival  Interceptor     Malawi        5  86.2
#> 82  Mean of log transformed holed area  Interceptor     Malawi        6   4.6
#> 83    SD of log transformed holed area  Interceptor     Malawi        6   1.9
#> 84                            Survival  Interceptor     Malawi        6  57.9
#> 85  Mean of log transformed holed area      LifeNet    Senegal        1   0.4
#> 86    SD of log transformed holed area      LifeNet    Senegal        1   1.1
#> 87                            Survival      LifeNet    Senegal        1 100.0
#> 88  Mean of log transformed holed area      LifeNet    Senegal        2   0.1
#> 89    SD of log transformed holed area      LifeNet    Senegal        2   0.6
#> 90                            Survival      LifeNet    Senegal        2  99.4
#> 91  Mean of log transformed holed area      LifeNet    Senegal        3   0.1
#> 92    SD of log transformed holed area      LifeNet    Senegal        3   0.6
#> 93                            Survival      LifeNet    Senegal        3  98.2
#> 94  Mean of log transformed holed area      LifeNet    Senegal        4   0.3
#> 95    SD of log transformed holed area      LifeNet    Senegal        4   0.7
#> 96                            Survival      LifeNet    Senegal        4  98.2
#> 97  Mean of log transformed holed area      LifeNet    Senegal        5   0.6
#> 98    SD of log transformed holed area      LifeNet    Senegal        5   0.8
#> 99                            Survival      LifeNet    Senegal        5  98.1
#> 100 Mean of log transformed holed area      LifeNet    Senegal        6   1.7
#> 101   SD of log transformed holed area      LifeNet    Senegal        6   2.3
#> 102                           Survival      LifeNet    Senegal        6  98.0
#> 103 Mean of log transformed holed area   Netprotect      Kenya        1   1.3
#> 104   SD of log transformed holed area   Netprotect      Kenya        1   2.7
#> 105                           Survival   Netprotect      Kenya        1 100.0
#> 106 Mean of log transformed holed area   Netprotect      Kenya        2   2.4
#> 107   SD of log transformed holed area   Netprotect      Kenya        2   3.3
#> 108                           Survival   Netprotect      Kenya        2 100.0
#> 109 Mean of log transformed holed area   Netprotect      Kenya        3   2.2
#> 110   SD of log transformed holed area   Netprotect      Kenya        3   2.4
#> 111                           Survival   Netprotect      Kenya        3  99.8
#> 112 Mean of log transformed holed area   Netprotect      Kenya        4   2.7
#> 113   SD of log transformed holed area   Netprotect      Kenya        4   1.8
#> 114                           Survival   Netprotect      Kenya        4  99.0
#> 115 Mean of log transformed holed area   Netprotect      Kenya        5   2.8
#> 116   SD of log transformed holed area   Netprotect      Kenya        5   2.2
#> 117                           Survival   Netprotect      Kenya        5  99.0
#> 118 Mean of log transformed holed area   Netprotect      Kenya        6   4.7
#> 119   SD of log transformed holed area   Netprotect      Kenya        6   2.9
#> 120                           Survival   Netprotect      Kenya        6  97.6
#> 121 Mean of log transformed holed area   Netprotect      Kenya        7   5.4
#> 122   SD of log transformed holed area   Netprotect      Kenya        7   2.4
#> 123                           Survival   Netprotect      Kenya        7  91.0
#> 124 Mean of log transformed holed area   Netprotect      Kenya        8   4.6
#> 125   SD of log transformed holed area   Netprotect      Kenya        8   3.1
#> 126                           Survival   Netprotect      Kenya        8  68.5
#> 127 Mean of log transformed holed area   Netprotect     Malawi        1   0.7
#> 128   SD of log transformed holed area   Netprotect     Malawi        1   1.4
#> 129                           Survival   Netprotect     Malawi        1  99.8
#> 130 Mean of log transformed holed area   Netprotect     Malawi        2   1.6
#> 131   SD of log transformed holed area   Netprotect     Malawi        2   1.3
#> 132                           Survival   Netprotect     Malawi        2  99.5
#> 133 Mean of log transformed holed area   Netprotect     Malawi        3   3.3
#> 134   SD of log transformed holed area   Netprotect     Malawi        3   2.1
#> 135                           Survival   Netprotect     Malawi        3  99.5
#> 136 Mean of log transformed holed area   Netprotect     Malawi        4   3.1
#> 137   SD of log transformed holed area   Netprotect     Malawi        4   1.8
#> 138                           Survival   Netprotect     Malawi        4  95.2
#> 139 Mean of log transformed holed area   Netprotect     Malawi        5   3.5
#> 140   SD of log transformed holed area   Netprotect     Malawi        5   1.8
#> 141                           Survival   Netprotect     Malawi        5  93.8
#> 142 Mean of log transformed holed area   Netprotect     Malawi        6   4.1
#> 143   SD of log transformed holed area   Netprotect     Malawi        6   2.1
#> 144                           Survival   Netprotect     Malawi        6  81.7
#> 145 Mean of log transformed holed area   Netprotect    Senegal        1   0.1
#> 146   SD of log transformed holed area   Netprotect    Senegal        1   0.7
#> 147                           Survival   Netprotect    Senegal        1 100.0
#> 148 Mean of log transformed holed area   Netprotect    Senegal        2   0.3
#> 149   SD of log transformed holed area   Netprotect    Senegal        2   1.2
#> 150                           Survival   Netprotect    Senegal        2  99.4
#> 151 Mean of log transformed holed area   Netprotect    Senegal        3   0.6
#> 152   SD of log transformed holed area   Netprotect    Senegal        3   1.5
#> 153                           Survival   Netprotect    Senegal        3  98.4
#> 154 Mean of log transformed holed area   Netprotect    Senegal        4   1.1
#> 155   SD of log transformed holed area   Netprotect    Senegal        4   2.2
#> 156                           Survival   Netprotect    Senegal        4  98.4
#> 157 Mean of log transformed holed area   Netprotect    Senegal        5   0.4
#> 158   SD of log transformed holed area   Netprotect    Senegal        5   1.2
#> 159                           Survival   Netprotect    Senegal        5  94.7
#> 160 Mean of log transformed holed area   Netprotect    Senegal        6   1.9
#> 161   SD of log transformed holed area   Netprotect    Senegal        6   3.5
#> 162                           Survival   Netprotect    Senegal        6  94.6
#> 163 Mean of log transformed holed area       Olyset      Benin        1   2.0
#> 164   SD of log transformed holed area       Olyset      Benin        1   2.5
#> 165                           Survival       Olyset      Benin        1  98.9
#> 166 Mean of log transformed holed area       Olyset      Benin        2   3.5
#> 167   SD of log transformed holed area       Olyset      Benin        2   2.8
#> 168                           Survival       Olyset      Benin        2  89.5
#> 169 Mean of log transformed holed area       Olyset      Benin        3   4.0
#> 170   SD of log transformed holed area       Olyset      Benin        3   2.9
#> 171                           Survival       Olyset      Benin        3  81.0
#> 172 Mean of log transformed holed area       Olyset      Benin        4   4.1
#> 173   SD of log transformed holed area       Olyset      Benin        4   2.8
#> 174                           Survival       Olyset      Benin        4  68.6
#> 175 Mean of log transformed holed area       Olyset      Kenya        1   1.7
#> 176   SD of log transformed holed area       Olyset      Kenya        1   2.4
#> 177                           Survival       Olyset      Kenya        1  99.9
#> 178 Mean of log transformed holed area       Olyset      Kenya        2   1.3
#> 179   SD of log transformed holed area       Olyset      Kenya        2   2.2
#> 180                           Survival       Olyset      Kenya        2  99.7
#> 181 Mean of log transformed holed area       Olyset      Kenya        3   3.0
#> 182   SD of log transformed holed area       Olyset      Kenya        3   2.2
#> 183                           Survival       Olyset      Kenya        3  97.2
#> 184 Mean of log transformed holed area       Olyset      Kenya        4   3.9
#> 185   SD of log transformed holed area       Olyset      Kenya        4   2.4
#> 186                           Survival       Olyset      Kenya        4  97.2
#> 187 Mean of log transformed holed area       Olyset      Kenya        5   4.5
#> 188   SD of log transformed holed area       Olyset      Kenya        5   3.5
#> 189                           Survival       Olyset      Kenya        5  93.4
#> 190 Mean of log transformed holed area       Olyset      Kenya        6   4.7
#> 191   SD of log transformed holed area       Olyset      Kenya        6   1.7
#> 192                           Survival       Olyset      Kenya        6  91.2
#> 193 Mean of log transformed holed area       Olyset      Kenya        7   5.5
#> 194   SD of log transformed holed area       Olyset      Kenya        7   2.8
#> 195                           Survival       Olyset      Kenya        7  85.9
#> 196 Mean of log transformed holed area       Olyset      Kenya        8   4.8
#> 197   SD of log transformed holed area       Olyset      Kenya        8   2.0
#> 198                           Survival       Olyset      Kenya        8  81.6
#> 199 Mean of log transformed holed area       Olyset     Malawi        1   2.2
#> 200   SD of log transformed holed area       Olyset     Malawi        1   2.2
#> 201                           Survival       Olyset     Malawi        1  99.6
#> 202 Mean of log transformed holed area       Olyset     Malawi        2   2.9
#> 203   SD of log transformed holed area       Olyset     Malawi        2   2.6
#> 204                           Survival       Olyset     Malawi        2  99.5
#> 205 Mean of log transformed holed area       Olyset     Malawi        3   4.5
#> 206   SD of log transformed holed area       Olyset     Malawi        3   2.4
#> 207                           Survival       Olyset     Malawi        3  98.0
#> 208 Mean of log transformed holed area       Olyset     Malawi        4   5.8
#> 209   SD of log transformed holed area       Olyset     Malawi        4   2.8
#> 210                           Survival       Olyset     Malawi        4  95.0
#> 211 Mean of log transformed holed area       Olyset     Malawi        5   5.6
#> 212   SD of log transformed holed area       Olyset     Malawi        5   1.8
#> 213                           Survival       Olyset     Malawi        5  89.8
#> 214 Mean of log transformed holed area       Olyset     Malawi        6   5.8
#> 215   SD of log transformed holed area       Olyset     Malawi        6   1.8
#> 216                           Survival       Olyset     Malawi        6  70.4
#> 217 Mean of log transformed holed area       Olyset Mozambique        2   4.8
#> 218   SD of log transformed holed area       Olyset Mozambique        2   2.0
#> 219                           Survival       Olyset Mozambique        2  98.0
#> 220 Mean of log transformed holed area       Olyset Mozambique        4   5.3
#> 221   SD of log transformed holed area       Olyset Mozambique        4   2.3
#> 222                           Survival       Olyset Mozambique        4  94.7
#> 223 Mean of log transformed holed area       Olyset Mozambique        6   7.1
#> 224   SD of log transformed holed area       Olyset Mozambique        6   2.3
#> 225                           Survival       Olyset Mozambique        6  84.6
#> 226 Mean of log transformed holed area       Olyset    Senegal        1   0.7
#> 227   SD of log transformed holed area       Olyset    Senegal        1   1.6
#> 228                           Survival       Olyset    Senegal        1 100.0
#> 229 Mean of log transformed holed area       Olyset    Senegal        2   1.4
#> 230   SD of log transformed holed area       Olyset    Senegal        2   2.2
#> 231                           Survival       Olyset    Senegal        2  99.4
#> 232 Mean of log transformed holed area       Olyset    Senegal        3   1.6
#> 233   SD of log transformed holed area       Olyset    Senegal        3   2.3
#> 234                           Survival       Olyset    Senegal        3  98.2
#> 235 Mean of log transformed holed area       Olyset    Senegal        4   3.5
#> 236   SD of log transformed holed area       Olyset    Senegal        4   2.7
#> 237                           Survival       Olyset    Senegal        4  92.3
#> 238 Mean of log transformed holed area       Olyset    Senegal        5   3.0
#> 239   SD of log transformed holed area       Olyset    Senegal        5   3.0
#> 240                           Survival       Olyset    Senegal        5  82.7
#> 241 Mean of log transformed holed area       Olyset    Senegal        6   4.1
#> 242   SD of log transformed holed area       Olyset    Senegal        6   2.8
#> 243                           Survival       Olyset    Senegal        6  76.0
#> 244 Mean of log transformed holed area       Olyset     Zambia        2   2.4
#> 245   SD of log transformed holed area       Olyset     Zambia        2   2.3
#> 246                           Survival       Olyset     Zambia        2 100.0
#> 247 Mean of log transformed holed area       Olyset     Zambia        3   3.2
#> 248   SD of log transformed holed area       Olyset     Zambia        3   2.3
#> 249                           Survival       Olyset     Zambia        3  89.8
#> 250 Mean of log transformed holed area       Olyset     Zambia        4   3.9
#> 251   SD of log transformed holed area       Olyset     Zambia        4   2.4
#> 252                           Survival       Olyset     Zambia        4  77.3
#> 253 Mean of log transformed holed area       Olyset     Zambia        5   4.5
#> 254   SD of log transformed holed area       Olyset     Zambia        5   2.2
#> 255                           Survival       Olyset     Zambia        5  57.3
#> 256 Mean of log transformed holed area PermaNet 2.0     Angola        2   3.3
#> 257   SD of log transformed holed area PermaNet 2.0     Angola        2   1.9
#> 258 Mean of log transformed holed area PermaNet 2.0     Angola        4   3.7
#> 259   SD of log transformed holed area PermaNet 2.0     Angola        4   2.0
#> 260 Mean of log transformed holed area PermaNet 2.0      Kenya        1   2.5
#> 261   SD of log transformed holed area PermaNet 2.0      Kenya        1   2.9
#> 262                           Survival PermaNet 2.0      Kenya        1  99.6
#> 263 Mean of log transformed holed area PermaNet 2.0      Kenya        2   1.2
#> 264   SD of log transformed holed area PermaNet 2.0      Kenya        2   2.3
#> 265                           Survival PermaNet 2.0      Kenya        2  98.5
#> 266 Mean of log transformed holed area PermaNet 2.0      Kenya        3   3.3
#> 267   SD of log transformed holed area PermaNet 2.0      Kenya        3   3.1
#> 268                           Survival PermaNet 2.0      Kenya        3  97.6
#> 269 Mean of log transformed holed area PermaNet 2.0      Kenya        4   3.1
#> 270   SD of log transformed holed area PermaNet 2.0      Kenya        4   2.9
#> 271                           Survival PermaNet 2.0      Kenya        4  97.1
#> 272 Mean of log transformed holed area PermaNet 2.0      Kenya        5   3.7
#> 273   SD of log transformed holed area PermaNet 2.0      Kenya        5   2.7
#> 274                           Survival PermaNet 2.0      Kenya        5  89.0
#> 275 Mean of log transformed holed area PermaNet 2.0      Kenya        6   5.1
#> 276   SD of log transformed holed area PermaNet 2.0      Kenya        6   3.4
#> 277                           Survival PermaNet 2.0      Kenya        6  81.1
#> 278 Mean of log transformed holed area PermaNet 2.0      Kenya        7   5.6
#> 279   SD of log transformed holed area PermaNet 2.0      Kenya        7   3.9
#> 280                           Survival PermaNet 2.0      Kenya        7  71.0
#> 281 Mean of log transformed holed area PermaNet 2.0      Kenya        8   3.9
#> 282   SD of log transformed holed area PermaNet 2.0      Kenya        8   3.4
#> 283                           Survival PermaNet 2.0      Kenya        8  58.9
#> 284 Mean of log transformed holed area PermaNet 2.0     Malawi        1   0.7
#> 285   SD of log transformed holed area PermaNet 2.0     Malawi        1   1.7
#> 286                           Survival PermaNet 2.0     Malawi        1 100.0
#> 287 Mean of log transformed holed area PermaNet 2.0     Malawi        2   1.7
#> 288   SD of log transformed holed area PermaNet 2.0     Malawi        2   2.4
#> 289                           Survival PermaNet 2.0     Malawi        2 100.0
#> 290 Mean of log transformed holed area PermaNet 2.0     Malawi        3   1.7
#> 291   SD of log transformed holed area PermaNet 2.0     Malawi        3   2.0
#> 292                           Survival PermaNet 2.0     Malawi        3  97.4
#> 293 Mean of log transformed holed area PermaNet 2.0     Malawi        4   2.7
#> 294   SD of log transformed holed area PermaNet 2.0     Malawi        4   2.3
#> 295                           Survival PermaNet 2.0     Malawi        4  94.6
#> 296 Mean of log transformed holed area PermaNet 2.0     Malawi        5   3.9
#> 297   SD of log transformed holed area PermaNet 2.0     Malawi        5   1.9
#> 298                           Survival PermaNet 2.0     Malawi        5  82.5
#> 299 Mean of log transformed holed area PermaNet 2.0     Malawi        6   5.3
#> 300   SD of log transformed holed area PermaNet 2.0     Malawi        6   2.3
#> 301                           Survival PermaNet 2.0     Malawi        6  55.8
#> 302 Mean of log transformed holed area PermaNet 2.0 Mozambique        2   1.6
#> 303   SD of log transformed holed area PermaNet 2.0 Mozambique        2   1.8
#> 304                           Survival PermaNet 2.0 Mozambique        2 100.0
#> 305 Mean of log transformed holed area PermaNet 2.0 Mozambique        4   3.3
#> 306   SD of log transformed holed area PermaNet 2.0 Mozambique        4   1.8
#> 307                           Survival PermaNet 2.0 Mozambique        4  95.2
#> 308 Mean of log transformed holed area PermaNet 2.0 Mozambique        6   5.0
#> 309   SD of log transformed holed area PermaNet 2.0 Mozambique        6   1.8
#> 310                           Survival PermaNet 2.0 Mozambique        6  77.2
#> 311 Mean of log transformed holed area PermaNet 2.0    Senegal        1   0.2
#> 312   SD of log transformed holed area PermaNet 2.0    Senegal        1   0.7
#> 313                           Survival PermaNet 2.0    Senegal        1  99.8
#> 314 Mean of log transformed holed area PermaNet 2.0    Senegal        2   0.5
#> 315   SD of log transformed holed area PermaNet 2.0    Senegal        2   1.4
#> 316                           Survival PermaNet 2.0    Senegal        2  99.4
#> 317 Mean of log transformed holed area PermaNet 2.0    Senegal        3   1.1
#> 318   SD of log transformed holed area PermaNet 2.0    Senegal        3   2.0
#> 319                           Survival PermaNet 2.0    Senegal        3  98.5
#> 320 Mean of log transformed holed area PermaNet 2.0    Senegal        4   1.5
#> 321   SD of log transformed holed area PermaNet 2.0    Senegal        4   2.4
#> 322                           Survival PermaNet 2.0    Senegal        4  98.5
#> 323 Mean of log transformed holed area PermaNet 2.0    Senegal        5   1.6
#> 324   SD of log transformed holed area PermaNet 2.0    Senegal        5   2.6
#> 325                           Survival PermaNet 2.0    Senegal        5  97.8
#> 326 Mean of log transformed holed area PermaNet 2.0    Senegal        6   3.3
#> 327   SD of log transformed holed area PermaNet 2.0    Senegal        6   3.3
#> 328                           Survival PermaNet 2.0    Senegal        6  97.8
#> 329 Mean of log transformed holed area PermaNet 2.0     Zambia        2   2.1
#> 330   SD of log transformed holed area PermaNet 2.0     Zambia        2   2.2
#> 331                           Survival PermaNet 2.0     Zambia        2 100.0
#> 332 Mean of log transformed holed area PermaNet 2.0     Zambia        3   2.9
#> 333   SD of log transformed holed area PermaNet 2.0     Zambia        3   2.2
#> 334                           Survival PermaNet 2.0     Zambia        3  94.2
#> 335 Mean of log transformed holed area PermaNet 2.0     Zambia        4   3.9
#> 336   SD of log transformed holed area PermaNet 2.0     Zambia        4   2.3
#> 337                           Survival PermaNet 2.0     Zambia        4  81.8
#> 338 Mean of log transformed holed area PermaNet 2.0     Zambia        5   4.3
#> 339   SD of log transformed holed area PermaNet 2.0     Zambia        5   2.4
#> 340                           Survival PermaNet 2.0     Zambia        5  66.2
#> 341 Mean of log transformed holed area PermaNet 3.0      Kenya        1   0.7
#> 342   SD of log transformed holed area PermaNet 3.0      Kenya        1   1.2
#> 343                           Survival PermaNet 3.0      Kenya        1 100.0
#> 344 Mean of log transformed holed area PermaNet 3.0      Kenya        2   1.5
#> 345   SD of log transformed holed area PermaNet 3.0      Kenya        2   2.1
#> 346                           Survival PermaNet 3.0      Kenya        2  99.8
#> 347 Mean of log transformed holed area PermaNet 3.0      Kenya        3   2.2
#> 348   SD of log transformed holed area PermaNet 3.0      Kenya        3   2.5
#> 349                           Survival PermaNet 3.0      Kenya        3  99.8
#> 350 Mean of log transformed holed area PermaNet 3.0      Kenya        4   1.4
#> 351   SD of log transformed holed area PermaNet 3.0      Kenya        4   2.1
#> 352                           Survival PermaNet 3.0      Kenya        4  99.1
#> 353 Mean of log transformed holed area PermaNet 3.0      Kenya        5   2.4
#> 354   SD of log transformed holed area PermaNet 3.0      Kenya        5   2.1
#> 355                           Survival PermaNet 3.0      Kenya        5  96.8
#> 356 Mean of log transformed holed area PermaNet 3.0      Kenya        6   3.6
#> 357   SD of log transformed holed area PermaNet 3.0      Kenya        6   2.4
#> 358                           Survival PermaNet 3.0      Kenya        6  95.1
#> 359 Mean of log transformed holed area PermaNet 3.0      Kenya        7   4.6
#> 360   SD of log transformed holed area PermaNet 3.0      Kenya        7   3.3
#> 361                           Survival PermaNet 3.0      Kenya        7  91.2
#> 362 Mean of log transformed holed area PermaNet 3.0      Kenya        8   3.3
#> 363   SD of log transformed holed area PermaNet 3.0      Kenya        8   2.4
#> 364                           Survival PermaNet 3.0      Kenya        8  88.5
#> 365                           Survival DawaPlus 2.0     Angola        2  90.2
#> 366                           Survival DawaPlus 2.0     Angola        4  74.9
#> 367 Mean of log transformed holed area DawaPlus 2.0      Kenya        1   2.4
#> 368   SD of log transformed holed area DawaPlus 2.0      Kenya        1   3.8
#> 369                           Survival DawaPlus 2.0      Kenya        1 100.0
#> 370 Mean of log transformed holed area DawaPlus 2.0      Kenya        2   1.5
#> 371   SD of log transformed holed area DawaPlus 2.0      Kenya        2   1.7
#> 372                           Survival DawaPlus 2.0      Kenya        2  99.2
#> 373 Mean of log transformed holed area DawaPlus 2.0      Kenya        3   1.8
#> 374   SD of log transformed holed area DawaPlus 2.0      Kenya        3   1.9
#> 375                           Survival DawaPlus 2.0      Kenya        3  97.3
#> 376 Mean of log transformed holed area DawaPlus 2.0      Kenya        4   2.1
#> 377   SD of log transformed holed area DawaPlus 2.0      Kenya        4   2.3
#> 378                           Survival DawaPlus 2.0      Kenya        4  94.8
#> 379 Mean of log transformed holed area DawaPlus 2.0      Kenya        5   3.9
#> 380   SD of log transformed holed area DawaPlus 2.0      Kenya        5   3.6
#> 381                           Survival DawaPlus 2.0      Kenya        5  91.6
#> 382 Mean of log transformed holed area DawaPlus 2.0      Kenya        6   4.7
#> 383   SD of log transformed holed area DawaPlus 2.0      Kenya        6   3.5
#> 384                           Survival DawaPlus 2.0      Kenya        6  87.3
#> 385 Mean of log transformed holed area DawaPlus 2.0      Kenya        7   5.3
#> 386   SD of log transformed holed area DawaPlus 2.0      Kenya        7   3.6
#> 387                           Survival DawaPlus 2.0      Kenya        7  77.4
#> 388 Mean of log transformed holed area DawaPlus 2.0      Kenya        8   6.4
#> 389   SD of log transformed holed area DawaPlus 2.0      Kenya        8   4.2
#> 390                           Survival DawaPlus 2.0      Kenya        8  66.7
#> 391 Mean of log transformed holed area DawaPlus 2.0    Senegal        1   0.2
#> 392   SD of log transformed holed area DawaPlus 2.0    Senegal        1   1.0
#> 393                           Survival DawaPlus 2.0    Senegal        1 100.0
#> 394 Mean of log transformed holed area DawaPlus 2.0    Senegal        2   0.5
#> 395   SD of log transformed holed area DawaPlus 2.0    Senegal        2   1.6
#> 396                           Survival DawaPlus 2.0    Senegal        2  98.7
#> 397 Mean of log transformed holed area DawaPlus 2.0    Senegal        3   1.3
#> 398   SD of log transformed holed area DawaPlus 2.0    Senegal        3   2.1
#> 399                           Survival DawaPlus 2.0    Senegal        3  98.2
#> 400 Mean of log transformed holed area DawaPlus 2.0    Senegal        4   1.8
#> 401   SD of log transformed holed area DawaPlus 2.0    Senegal        4   2.6
#> 402                           Survival DawaPlus 2.0    Senegal        4  97.5
#> 403 Mean of log transformed holed area DawaPlus 2.0    Senegal        5   1.5
#> 404   SD of log transformed holed area DawaPlus 2.0    Senegal        5   2.5
#> 405                           Survival DawaPlus 2.0    Senegal        5  97.4
#> 406 Mean of log transformed holed area DawaPlus 2.0    Senegal        6   0.9
#> 407   SD of log transformed holed area DawaPlus 2.0    Senegal        6   1.8
#> 408                           Survival DawaPlus 2.0    Senegal        6  97.4
#>       L95   U95
#> 1    -0.1   0.3
#> 2     0.3   0.6
#> 3   100.0 100.0
#> 4     0.3   1.9
#> 5     1.5   2.8
#> 6    99.0 100.0
#> 7     1.2   2.9
#> 8     1.5   2.8
#> 9    99.0 100.0
#> 10    0.7   1.7
#> 11    1.0   1.8
#> 12   95.6  98.3
#> 13    0.6   2.1
#> 14    1.5   2.7
#> 15   95.6  98.3
#> 16    1.7   3.4
#> 17    2.1   3.4
#> 18   94.7  97.8
#> 19    2.6   4.2
#> 20    1.9   3.1
#> 21   92.1  96.2
#> 22    2.3   4.3
#> 23    2.3   3.8
#> 24   86.1  92.0
#> 25    0.1   2.0
#> 26    1.5   3.0
#> 27   99.9 100.0
#> 28    1.0   2.4
#> 29    1.1   2.1
#> 30   97.0  99.5
#> 31    1.0   2.9
#> 32    1.3   2.8
#> 33   94.4  98.1
#> 34    1.5   3.0
#> 35    1.3   2.4
#> 36   93.5  97.3
#> 37    1.7   3.5
#> 38    1.4   2.7
#> 39   84.4  90.9
#> 40    2.8   3.8
#> 41    1.4   2.1
#> 42   59.5  70.1
#> 43    0.4   2.3
#> 44    1.5   3.0
#> 45   99.9 100.0
#> 46    0.9   3.8
#> 47    2.0   4.4
#> 48   98.8  99.9
#> 49    1.1   2.8
#> 50    1.7   3.1
#> 51   98.8  99.9
#> 52    1.6   3.3
#> 53    1.7   3.1
#> 54   98.2  99.7
#> 55    1.9   4.2
#> 56    2.0   3.8
#> 57   96.2  98.8
#> 58    2.6   4.5
#> 59    2.1   3.6
#> 60   87.8  94.5
#> 61    4.4   7.2
#> 62    2.9   5.4
#> 63   67.2  77.6
#> 64    3.8   5.9
#> 65    2.7   4.5
#> 66   44.7  55.9
#> 67   -0.1   1.0
#> 68    0.9   1.8
#> 69   99.1 100.0
#> 70    1.3   3.0
#> 71    1.5   2.8
#> 72   98.4  99.9
#> 73    2.1   4.4
#> 74    2.0   3.6
#> 75   92.9  97.4
#> 76    2.4   4.3
#> 77    1.5   2.9
#> 78   85.6  91.9
#> 79    3.7   5.4
#> 80    1.7   2.9
#> 81   82.7  89.4
#> 82    4.1   5.1
#> 83    1.6   2.4
#> 84   52.2  63.2
#> 85    0.1   0.7
#> 86    0.9   1.3
#> 87   99.8 100.0
#> 88    0.0   0.2
#> 89    0.5   0.7
#> 90   98.1 100.0
#> 91    0.0   0.2
#> 92    0.5   0.7
#> 93   96.3  99.4
#> 94    0.0   0.6
#> 95    0.6   0.9
#> 96   96.3  99.4
#> 97   -2.4   3.3
#> 98    0.6   9.4
#> 99   96.0  99.4
#> 100  -1.2   4.6
#> 101   1.4   6.6
#> 102  95.3  99.4
#> 103  -0.3   2.9
#> 104   1.9   4.5
#> 105 100.0 100.0
#> 106   0.6   4.3
#> 107   2.3   5.5
#> 108  99.9 100.0
#> 109   1.1   3.3
#> 110   1.8   3.5
#> 111  99.1 100.0
#> 112   1.9   3.5
#> 113   1.4   2.6
#> 114  97.9  99.7
#> 115   1.7   4.0
#> 116   1.6   3.4
#> 117  97.9  99.7
#> 118   3.7   5.8
#> 119   2.3   4.0
#> 120  95.9  98.9
#> 121   4.3   6.6
#> 122   1.7   3.8
#> 123  87.6  94.0
#> 124   3.3   5.9
#> 125   2.4   4.3
#> 126  62.9  74.2
#> 127   0.1   1.3
#> 128   1.1   2.0
#> 129  98.1 100.0
#> 130   0.8   2.5
#> 131   1.0   2.2
#> 132  97.7 100.0
#> 133   2.1   4.6
#> 134   1.5   3.4
#> 135  97.7 100.0
#> 136   2.4   3.8
#> 137   1.4   2.5
#> 138  93.1  96.9
#> 139   2.7   4.2
#> 140   1.4   2.6
#> 141  91.4  95.8
#> 142   3.6   4.5
#> 143   1.8   2.5
#> 144  77.4  85.6
#> 145   0.0   0.2
#> 146   0.6   0.7
#> 147  99.9 100.0
#> 148   0.1   0.5
#> 149   1.0   1.3
#> 150  98.4  99.9
#> 151   0.3   0.9
#> 152   1.3   1.7
#> 153  96.9  99.5
#> 154   0.5   1.7
#> 155   1.8   2.7
#> 156  96.9  99.4
#> 157  -0.5   1.3
#> 158   0.8   2.3
#> 159  90.0  97.6
#> 160  -6.4   9.6
#> 161   1.8  19.9
#> 162  89.9  97.6
#> 163   1.7   2.4
#> 164   2.2   2.7
#> 165  98.4  99.4
#> 166   3.1   3.8
#> 167   2.5   3.1
#> 168  87.9  91.1
#> 169   3.6   4.4
#> 170   2.6   3.2
#> 171  78.9  83.0
#> 172   3.7   4.4
#> 173   2.5   3.1
#> 174  65.9  71.2
#> 175   0.6   2.8
#> 176   1.8   3.6
#> 177  99.1 100.0
#> 178   0.2   2.5
#> 179   1.6   3.5
#> 180  98.8 100.0
#> 181   2.0   4.1
#> 182   1.7   3.2
#> 183  95.5  98.6
#> 184   2.7   5.1
#> 185   1.8   3.6
#> 186  95.5  98.6
#> 187   2.4   6.8
#> 188   2.4   6.2
#> 189  90.8  95.7
#> 190   3.9   5.5
#> 191   1.3   2.5
#> 192  88.2  93.9
#> 193   4.2   6.9
#> 194   2.0   4.3
#> 195  82.1  89.5
#> 196   3.7   5.9
#> 197   1.4   3.2
#> 198  77.1  86.0
#> 199   1.1   3.2
#> 200   1.7   3.2
#> 201  98.9 100.0
#> 202   1.4   4.3
#> 203   1.9   4.2
#> 204  98.7  99.9
#> 205   3.2   5.7
#> 206   1.8   3.7
#> 207  96.3  99.1
#> 208   4.4   7.3
#> 209   2.1   4.3
#> 210  92.8  96.9
#> 211   4.7   6.5
#> 212   1.4   2.7
#> 213  86.7  92.5
#> 214   5.3   6.3
#> 215   1.5   2.2
#> 216  65.7  74.8
#> 217   4.0   5.7
#> 218   1.5   2.8
#> 219  94.3 100.0
#> 220   3.9   6.7
#> 221   1.7   3.9
#> 222  88.7  98.5
#> 223   5.8   8.4
#> 224   1.6   3.8
#> 225  74.7  92.2
#> 226   0.6   0.8
#> 227   1.5   1.7
#> 228  99.9 100.0
#> 229   1.2   1.7
#> 230   2.1   2.4
#> 231  98.7  99.8
#> 232   1.4   1.8
#> 233   2.2   2.5
#> 234  97.0  99.2
#> 235   3.2   3.8
#> 236   2.5   2.9
#> 237  89.9  94.5
#> 238   2.6   3.3
#> 239   2.7   3.2
#> 240  79.2  86.0
#> 241   3.6   4.5
#> 242   2.6   3.2
#> 243  71.3  80.2
#> 244   2.2   2.6
#> 245   2.2   2.5
#> 246  99.9 100.0
#> 247   3.0   3.4
#> 248   2.1   2.4
#> 249  86.7  92.6
#> 250   3.7   4.2
#> 251   2.3   2.6
#> 252  72.8  81.5
#> 253   4.3   4.8
#> 254   2.1   2.4
#> 255  51.9  62.7
#> 256   2.8   3.8
#> 257   1.6   2.3
#> 258   3.2   4.1
#> 259   1.7   2.4
#> 260   0.7   4.5
#> 261   2.0   5.1
#> 262  99.0 100.0
#> 263   0.1   2.4
#> 264   1.7   3.6
#> 265  97.3  99.3
#> 266   2.0   4.7
#> 267   2.3   4.5
#> 268  96.3  98.8
#> 269   1.7   4.5
#> 270   2.2   4.3
#> 271  95.5  98.4
#> 272   2.4   5.1
#> 273   2.0   4.0
#> 274  85.8  92.0
#> 275   3.9   6.5
#> 276   2.6   4.9
#> 277  77.1  84.8
#> 278   4.0   7.6
#> 279   2.8   6.1
#> 280  66.1  75.7
#> 281   2.6   5.3
#> 282   2.6   4.9
#> 283  52.7  64.8
#> 284  -0.1   1.6
#> 285   1.2   2.5
#> 286 100.0 100.0
#> 287   0.2   3.3
#> 288   1.7   4.1
#> 289  99.7 100.0
#> 290   0.7   2.7
#> 291   1.5   3.1
#> 292  95.2  99.0
#> 293   1.5   3.8
#> 294   1.7   3.5
#> 295  91.9  96.8
#> 296   3.1   4.7
#> 297   1.5   2.7
#> 298  78.1  86.5
#> 299   4.6   5.9
#> 300   1.9   2.9
#> 301  50.3  61.5
#> 302   1.1   2.2
#> 303   1.5   2.2
#> 304  99.6 100.0
#> 305   2.7   3.9
#> 306   1.5   2.4
#> 307  90.8  98.0
#> 308   4.3   5.6
#> 309   1.5   2.4
#> 310  69.6  84.1
#> 311   0.1   0.3
#> 312   0.7   0.8
#> 313  99.5 100.0
#> 314   0.3   0.7
#> 315   1.3   1.6
#> 316  98.6  99.9
#> 317   0.8   1.4
#> 318   1.8   2.3
#> 319  97.2  99.4
#> 320   1.0   1.9
#> 321   2.1   2.7
#> 322  97.2  99.4
#> 323   0.6   2.7
#> 324   2.1   3.6
#> 325  95.7  99.0
#> 326   1.1   5.5
#> 327   2.3   5.8
#> 328  95.5  99.0
#> 329   1.9   2.3
#> 330   2.0   2.3
#> 331 100.0 100.0
#> 332   2.7   3.1
#> 333   2.1   2.3
#> 334  91.6  96.5
#> 335   3.7   4.2
#> 336   2.1   2.4
#> 337  77.7  85.5
#> 338   4.0   4.7
#> 339   2.2   2.6
#> 340  60.5  71.5
#> 341   0.2   1.2
#> 342   0.9   1.7
#> 343  99.9 100.0
#> 344   0.5   2.5
#> 345   1.5   3.3
#> 346  99.3 100.0
#> 347   1.1   3.4
#> 348   1.9   3.6
#> 349  99.3 100.0
#> 350   0.5   2.2
#> 351   1.6   3.0
#> 352  98.1  99.8
#> 353   1.5   3.3
#> 354   1.6   3.0
#> 355  95.0  98.2
#> 356   2.8   4.5
#> 357   2.0   3.3
#> 358  92.9  96.9
#> 359   3.3   6.0
#> 360   2.5   4.8
#> 361  88.3  93.8
#> 362   2.6   4.1
#> 363   1.9   3.2
#> 364  85.0  91.6
#> 365  85.0  94.3
#> 366  68.2  81.3
#> 367   0.5   4.5
#> 368   2.7   6.4
#> 369  99.9 100.0
#> 370   0.7   2.3
#> 371   1.3   2.6
#> 372  98.2  99.8
#> 373   1.0   2.7
#> 374   1.5   2.9
#> 375  95.7  98.6
#> 376   1.2   3.2
#> 377   1.7   3.4
#> 378  92.7  96.7
#> 379   2.2   5.9
#> 380   2.6   6.0
#> 381  88.9  94.1
#> 382   3.2   6.7
#> 383   2.5   5.7
#> 384  84.0  90.3
#> 385   3.8   7.3
#> 386   2.5   5.9
#> 387  72.7  81.8
#> 388   4.6   9.1
#> 389   3.0   7.2
#> 390  61.2  72.0
#> 391   0.1   0.4
#> 392   0.9   1.1
#> 393  99.8 100.0
#> 394   0.3   0.8
#> 395   1.4   1.8
#> 396  97.4  99.6
#> 397   0.9   1.6
#> 398   1.9   2.4
#> 399  96.6  99.3
#> 400   1.3   2.2
#> 401   2.3   3.0
#> 402  95.7  98.9
#> 403   0.5   2.6
#> 404   1.9   3.6
#> 405  95.5  98.8
#> 406  -0.5   2.5
#> 407   1.2   3.6
#> 408  95.2  98.8
#> 
#> 
#> $IRS_params
#>    Parameterisation Measured_quantity  Decay_type Initial_effect     L    k
#> 1             IRS01            alphai        step          0.336 0.500   NA
#> 2             IRS02            alphai     Weibull         -0.030 0.270 1.55
#> 3             IRS02               PBi     Weibull          0.160 0.270 1.55
#> 4             IRS02               PDi     Weibull          0.550 0.270 1.55
#> 5             IRS03            alphai     Weibull         -0.230 0.448 1.11
#> 6             IRS03               PBi     Weibull          0.120 0.448 1.11
#> 7             IRS03               PDi     Weibull          0.990 0.448 1.11
#> 8             IRS04            alphai     Weibull         -0.240 0.299 1.51
#> 9             IRS04               PBi     Weibull          0.120 0.299 1.51
#> 10            IRS04               PDi     Weibull          0.990 0.299 1.51
#> 11            IRS05               PDi exponential          0.810 0.170   NA
#> 12            IRS06            alphai   approxfun          1.000    NA   NA
#> 13            IRS06               PBi   approxfun          1.000    NA   NA
#> 14            IRS06               PCi   approxfun          1.000    NA   NA
#> 15            IRS08            alphai   approxfun          1.000    NA   NA
#> 16            IRS08               PBi   approxfun          1.000    NA   NA
#> 17            IRS08               PCi   approxfun          1.000    NA   NA
#> 18            IRS10            alphai exponential          0.560 0.500   NA
#> 19            IRS10               PDi exponential          0.240 0.500   NA
#> 20            IRS12            alphai   approxfun          1.000    NA   NA
#> 21            IRS12               PBi   approxfun          1.000    NA   NA
#> 22            IRS12               PCi   approxfun          1.000    NA   NA
#> 23            IRS14            alphai   approxfun          1.000    NA   NA
#> 24            IRS14               PBi   approxfun          1.000    NA   NA
#> 25            IRS14               PCi   approxfun          1.000    NA   NA
#> 26            IRS16            alphai exponential          0.310 0.250   NA
#> 27            IRS16               PDi exponential          0.770 0.250   NA
#> 28            IRS17            alphai   approxfun          1.000    NA   NA
#> 29            IRS17               PBi   approxfun          1.000    NA   NA
#> 30            IRS17               PCi   approxfun          1.000    NA   NA
#> 31            IRS18            alphai   approxfun          1.000    NA   NA
#> 32            IRS18               PBi   approxfun          1.000    NA   NA
#> 33            IRS18               PCi   approxfun          1.000    NA   NA
#>    T_0.02 T_0.08 T_0.15 T_0.23 T_0.29 T_0.46 T_0.041666667     T_0.125
#> 1      NA     NA     NA     NA     NA     NA            NA          NA
#> 2      NA     NA     NA     NA     NA     NA            NA          NA
#> 3      NA     NA     NA     NA     NA     NA            NA          NA
#> 4      NA     NA     NA     NA     NA     NA            NA          NA
#> 5      NA     NA     NA     NA     NA     NA            NA          NA
#> 6      NA     NA     NA     NA     NA     NA            NA          NA
#> 7      NA     NA     NA     NA     NA     NA            NA          NA
#> 8      NA     NA     NA     NA     NA     NA            NA          NA
#> 9      NA     NA     NA     NA     NA     NA            NA          NA
#> 10     NA     NA     NA     NA     NA     NA            NA          NA
#> 11     NA     NA     NA     NA     NA     NA            NA          NA
#> 12  -0.43  -0.36  -1.24  -0.40  -0.02      0            NA          NA
#> 13   0.50   0.47   0.70   0.27   0.00      0            NA          NA
#> 14   1.00   0.65   0.66   0.65   0.50      0            NA          NA
#> 15     NA     NA     NA     NA     NA     NA   -0.04671701 0.089762216
#> 16     NA     NA     NA     NA     NA     NA    0.05044087 0.011631425
#> 17     NA     NA     NA     NA     NA     NA    0.97400000 0.079500000
#> 18     NA     NA     NA     NA     NA     NA            NA          NA
#> 19     NA     NA     NA     NA     NA     NA            NA          NA
#> 20     NA     NA     NA     NA     NA     NA    0.36849137 0.513614537
#> 21     NA     NA     NA     NA     NA     NA    0.11131628 0.144058962
#> 22     NA     NA     NA     NA     NA     NA    0.80797947 0.696914111
#> 23  -0.21  -0.15  -0.98  -0.42  -0.41      0            NA          NA
#> 24   0.40   0.37   0.38   0.12   0.12      0            NA          NA
#> 25   0.30   0.20   0.44   0.58   0.24      0            NA          NA
#> 26     NA     NA     NA     NA     NA     NA            NA          NA
#> 27     NA     NA     NA     NA     NA     NA            NA          NA
#> 28     NA     NA     NA     NA     NA     NA   -0.03492146 0.034897358
#> 29     NA     NA     NA     NA     NA     NA    0.14820692 0.006880912
#> 30     NA     NA     NA     NA     NA     NA    0.92050000 0.093200000
#> 31     NA     NA     NA     NA     NA     NA            NA          NA
#> 32     NA     NA     NA     NA     NA     NA            NA          NA
#> 33     NA     NA     NA     NA     NA     NA            NA          NA
#>    T_0.208333333 T_0.291666667    T_0.375 T_0.458333333 T_0.541666667
#> 1             NA            NA         NA            NA            NA
#> 2             NA            NA         NA            NA            NA
#> 3             NA            NA         NA            NA            NA
#> 4             NA            NA         NA            NA            NA
#> 5             NA            NA         NA            NA            NA
#> 6             NA            NA         NA            NA            NA
#> 7             NA            NA         NA            NA            NA
#> 8             NA            NA         NA            NA            NA
#> 9             NA            NA         NA            NA            NA
#> 10            NA            NA         NA            NA            NA
#> 11            NA            NA         NA            NA            NA
#> 12            NA            NA         NA            NA            NA
#> 13            NA            NA         NA            NA            NA
#> 14            NA            NA         NA            NA            NA
#> 15   0.135581254            NA         NA  -0.052419018             0
#> 16   0.200402542            NA         NA   0.003163710             0
#> 17   0.515800000            NA         NA   0.043500000             0
#> 18            NA            NA         NA            NA            NA
#> 19            NA            NA         NA            NA            NA
#> 20   0.735871137    0.60789057 0.49881679   0.526661326             0
#> 21   0.124382577    0.01994285 0.02077001   0.025741281             0
#> 22   0.719230343    0.34244395 0.12063083   0.259375623             0
#> 23            NA            NA         NA            NA            NA
#> 24            NA            NA         NA            NA            NA
#> 25            NA            NA         NA            NA            NA
#> 26            NA            NA         NA            NA            NA
#> 27            NA            NA         NA            NA            NA
#> 28   0.034207189            NA         NA  -0.040995824             0
#> 29   0.006223192            NA         NA   0.002981936             0
#> 30   0.088700000            NA         NA   0.035300000             0
#> 31            NA            NA         NA            NA            NA
#> 32            NA            NA         NA            NA            NA
#> 33            NA            NA         NA            NA            NA
#>