Extract activity patterns of humans or mosquitoes

This function returns the activity patterns of humans or mosquitoes available in the package database based on a list of provided criteria.

list_activity(
  species = "Anopheles gambiae",
  sampling = "HBI",
  country = NULL,
  site = NULL
)

Arguments

species

string corresponding to the name of the species to extract the activity patterns for. For human activity patterns, this parameters should be "Homo Sapiens". For mosquito activity patterns, this name must match one of the Anopheles species provided in the package. The default value is "Anopheles gambiae". To see all available Anopheles species, use the function list_all_species().

sampling

can only be one of the following strings:

IND: proportion of humans indoors; only relevant for humans (species parameter is "Homo Sapiens").

BED: proportion of humans in bed; only relevant for humans (species parameter is "Homo Sapiens").

HBI: proportion of mosquito biting of humans occurring indoors; only relevant for mosquito species.

HBO: proportion of mosquito biting of humans occurring outdoors; only relevant for mosquito species. #'

ABO: proportion of mosquito biting of animals occurring outdoors; only relevant for mosquito species. #'

HB: proportion of mosquito biting of humans (indoors and outdoors); only relevant for mosquito species. The default value is "HBI".

country

name of the country. Default value is NULL. If the country is not provided for the provided species and sampling code, then data for all available countries will be displayed.

site

name of the geographical location within the country. Default value is NULL. If this parameter is not provided for the provided species and sampling code, then data for all available sites and countries will be displayed).

Value

data frame with the following columns:

ID: id of the database entry

species: name of the considered species

sampling: sampling type

country: country where measurements were taken

site: geographical site where measurements were taken

variable: time points; these are chronologically ordered in the output data frame

value: value of the sampling at each time point

Author

Monica Golumbeanu, monica.golumbeanu@swisstph.ch

Examples

# List the proportion of human biting indoors in the Rachuonyo region, Kenya.
humans_indoors = list_activity(species = "Homo sapiens", sampling = "IND",
site="Rachuonyo")
print(humans_indoors)
#>    id      species sampling country      site        hour      value
#> 17  2 Homo sapiens      IND   Kenya Rachuonyo 16.00_17.00 0.00000000
#> 18  2 Homo sapiens      IND   Kenya Rachuonyo 17.00_18.00 0.15566667
#> 19  2 Homo sapiens      IND   Kenya Rachuonyo 18.00_19.00 0.39518333
#> 20  2 Homo sapiens      IND   Kenya Rachuonyo 19.00_20.00 1.00000000
#> 21  2 Homo sapiens      IND   Kenya Rachuonyo 20.00_21.00 1.00000000
#> 22  2 Homo sapiens      IND   Kenya Rachuonyo 21.00_22.00 1.00000000
#> 23  2 Homo sapiens      IND   Kenya Rachuonyo 22.00_23.00 1.00000000
#> 24  2 Homo sapiens      IND   Kenya Rachuonyo 23.00_00.00 1.00000000
#> 25  2 Homo sapiens      IND   Kenya Rachuonyo 00.00_01.00 1.00000000
#> 26  2 Homo sapiens      IND   Kenya Rachuonyo 01.00_02.00 1.00000000
#> 27  2 Homo sapiens      IND   Kenya Rachuonyo 02.00_03.00 1.00000000
#> 28  2 Homo sapiens      IND   Kenya Rachuonyo 03.00_04.00 1.00000000
#> 29  2 Homo sapiens      IND   Kenya Rachuonyo 04.00_05.00 1.00000000
#> 30  2 Homo sapiens      IND   Kenya Rachuonyo 05.00_06.00 0.53900000
#> 31  2 Homo sapiens      IND   Kenya Rachuonyo 06.00_07.00 0.03533333
#> 32  2 Homo sapiens      IND   Kenya Rachuonyo 07.00_08.00 0.00000000