Print all the species whose entomological parameters are provided in the package database.

This function can be used to list all mosquito species whose entomological parameters are available in the package.

list_all_species()

Value

The function prints all the species names and also returns a vector containing them.

Author

Monica Golumbeanu, monica.golumbeanu@swisstph.ch

Examples

species_vec = list_all_species()
print(species_vec)
#>  [1] Anopheles arabiensis      Anopheles gambiae        
#>  [3] Anopheles funestus        Anopheles nili           
#>  [5] Anopheles moucheti        Anopheles melas          
#>  [7] Anopheles merus           Anopheles ovengensis     
#>  [9] Anopheles carnevalei      Anopheles albimanus      
#> [11] Anopheles darlingi        Anopheles aquasalis      
#> [13] Anopheles marajoara       Anopheles albitarsis E   
#> [15] Anopheles nuneztovari B/C Anopheles albitarsis A   
#> [17] Anopheles albitarsis B    Anopheles quadrimaculatus
#> [19] Anopheles freeborni       Anopheles aconitus       
#> [21] Anopheles maculatus       Anopheles barbirostris   
#> [23] Anopheles flavirostris    Anopheles sundaicus      
#> [25] Anopheles sinensis        Anopheles harrisoni      
#> [27] Anopheles dirus           Anopheles minimus        
#> [29] Anopheles balabacensis    Anopheles stephensi      
#> [31] Anopheles anthropophagus  Anopheles baimaii        
#> [33] Anopheles leucosphyrus    Anopheles cracens        
#> [35] Anopheles fluviatilis T   Anopheles culicifacies A 
#> [37] Anopheles culicifacies B  Anopheles culicifacies C 
#> [39] Anopheles fluviatilis S   Anopheles culicifacies D 
#> [41] Anopheles epiroticus      Anopheles koliensis      
#> [43] Anopheles latens          Anopheles subpictus B    
#> [45] Anopheles farauti         Anopheles punctulatus    
#> [47] Anopheles fluviatilis     Anopheles annularis A    
#> [49] Anopheles culicifacies E  Anopheles splendidus     
#> [51] Anopheles nivipes         Anopheles pseudojamesi   
#> [53] Anopheles jamesii         Anopheles kochi          
#> [55] Anopheles philippinensis  Anopheles tessellatus    
#> [57] Anopheles vagus           Gambiae                  
#> [59] Funestus                  Nili                     
#> [61] Pseudopunctipennis        Nuneztovari              
#> [63] Albitarsis                Quadrimaculatus          
#> [65] Annularis                 Subpictus                
#> [67] Dirus                     Maculatus                
#> [69] Barbirostris              Sundaicus                
#> [71] Hyrcanus                  Leucosphyrus             
#> [73] Punctulatus               Jamesii                  
#> [75] GENUS                    
#> 75 Levels: Albitarsis Annularis Anopheles aconitus ... Sundaicus