Vector-specific estimated entomological parameters

This data object contains bionomic parameterizations for 57 Anopheles mosquito species and 17 complexes (families of species). These parameters have been estimated using a Bayesian hierarchical model accounting for the phylogeny of the Anopheles genus and informed by entomological data available from the Malaria Atlas Project and other published field studies.

vec_ent_param

Format

A data frame with the following columns:

species_name: name of the mosquito species

M: parous rate (proportion of host-seeking mosquitoes that have laid eggs at least once)

Chi: human blood index (proportion of blood meals derived from humans by mosquitoes)

A0: sac rate (proportion of mosquitoes who laid eggs the same day)

zeta.3: relative availability of different non-human hosts

td: proportion of a day that a mosquito actively seeks a host

tau: time required for a mosquito that has encountered a host to return to host-seeking

ts: duration of the extrinsic incubation period (time required for sporozoites to develop to mosquitoes)

to: oocyst development time

endophily: proportion of indoor resting mosquitoes #'

endophagy: proportion of indoor feeding mosquitoes

Value

vector-specific entomological parameters

Author

Monica Golumbeanu, monica.golumbeanu@swisstph.ch

Examples

# Print the first lines of the data frame
head(vec_ent_param)
#>           species_name         M        M.sd       Chi        A0      A0.sd
#> 1 Anopheles arabiensis 0.5384622 0.003500333 0.4749833 0.5744970 0.01416144
#> 2    Anopheles gambiae 0.6134499 0.003819127 0.7947577 0.6419328 0.07280429
#> 3   Anopheles funestus 0.7565665 0.008446507 0.6243968 0.5899116 0.02053846
#> 4       Anopheles nili 0.6593646 0.006636148 0.6707206 0.4951913 0.05195866
#> 5   Anopheles moucheti 0.7268564 0.007056198 0.5828408 0.4951913 0.05195866
#> 6      Anopheles melas 0.6559240 0.001829826 0.5064544 0.5744970 0.01416144
#>   zeta.3   td tau ts to endophily endophily.sd endophagy endophagy.sd
#> 1      1 0.33   3 10  5 0.7743085   0.01233146 0.6578578  0.004550147
#> 2      1 0.33   3 10  5 0.7743085   0.01233146 0.5604133  0.005056590
#> 3      1 0.33   3 10  5 0.7424881   0.08777241 0.5343858  0.005546232
#> 4      1 0.33   3 10  5 0.6944887   0.13455023 0.3439954  0.024569634
#> 5      1 0.33   3 10  5 0.6944887   0.13455023 0.3996990  0.042303792
#> 6      1 0.33   3 10  5 0.7743085   0.01233146 0.4920965  0.002755030